Started adding Pester tests.
Basic testing of Protect-Data and Unprotect-Data is done.  Still need to add tests of other commands, and perhaps some internal functions.

Current certificate tests are done with a dynamically generated self-signed cert, which means all tests are done with the -SkipCertificateVerification parameter.
I would prefer to keep the tests portable so they can run anywhere, but I'm probably going to have to bite the bullet on that and hard-code the thumbprint of a
valid, trusted certificate on my computer instead.