Allow legacy (non-OAEP) padding for certificate encryption
As reported in issue #3 , some certificates may not work with OAEP padding.  Protect-Data and Add-ProtectedDataCredential now have a -UseLegacyPadding switch which allows you to encrypt data using PKCS#1 v1.5 padding instead.  This flag is saved along with the KeyData object, so there is no need to worry about the padding when decrypting.