Deprecating the -SkipCertificateVerification switch
The module now just behaves as though that switch were set to $True in all cases.

Reasoning is that by specifying a particular certificate, we're already implying that we trust it.  Going out to a CA or checking the validity period is redundant at that point (though the caller can take those steps before calling a ProtectedData command, if they wish.)