Automatic detection of decryption certificates
The Certificate parameter of Unprotect-Data is no longer mandatory.  If a matching certificate is found in the Personal store of either the user or the machine (and the private key is accessible in either location), then that cert will automatically be used to decrypt.

Also fixed a bug that was producing KeyData objects that were hashtables instead of PSCustomObjects when encrypting data with an ECDH certificate.  (All of the code still works either way, but it was inconsistent.)