Password hashing updates
Default password iteration count increased from 1000 to 50000.  This takes about a half second to has a single password on my computer, which is a much more reasonable standard from the point of view of making it hard to brute-force.

Also, found and fixed a bug in Add-ProtectedDataCredential, which has a -PasswordIterationCount parameter that was being ignored.  Tests have been updated to make sure that doesn't happen again.