Updated build script
So it'll stop crapping out trying to remove / overwrite Security.Cryptography.dll if I have the module imported already.  DLL is now signed right in the source control repo, and the build script no longer overwrites that signature, so its hash doesn't change and the build script won't try to copy it.